refactor: extract RankingImportService, batch calculateRankings, add Actuator, move newrelic.yml
- Extract importRankings() from ImportService into a new RankingImportService to
  eliminate the @Lazy self-injection AOP proxy workaround; ImportService now delegates
  to RankingImportService via normal constructor injection
- Batch-prefetch all 'overall' rankings for a gender in a single DB query inside
  calculateRankings(), then compute all derived age-group rankings in memory and
  persist with one saveAll() call - reduces DB round-trips from ~20 to 2 per import
- Remove redundant @EvictImportCaches from scanAndImport(); cache eviction now fires
  exactly once per successful import via RankingImportService
- Add spring-boot-starter-actuator; expose only /actuator/health and /actuator/info
  with no detail leak (show-details: never)
- Move newrelic.yml into newrelic/ subfolder and update Dockerfile COPY path so all
  New Relic artefacts live in one place
