+++ pom.xml
@@ -57,10 +57,14 @@
       <version>${jte.version}</version>
     </dependency>
 
-    <!-- Cache / Security -->
+    <!-- Cache / Security / Actuator -->
     <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-cache</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-actuator</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
